requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
openai==1.57.4
python-multipart==0.0.12
aiofiles==24.1.0
pandas==2.2.3
